Make loadMesh correctly resize buffer

+++ src/terrain.h
@@ -30,8 +30,8 @@ public:
     render();
 
 private:
-    static constexpr int granularity = 400;
-    static constexpr int iterations = 100;
+    static constexpr size_t granularity = 400;
+    static constexpr size_t iterations = 100;
 
     GLuint m_VBO, m_loadingVBO, m_IBO;
